feat:  views client component for counter and display views
