--- src/components/content/index.ts
@@ -1,1 +1,0 @@
-export { default as Mdx } from "./Mdx";
