--- src/components/icons/index.ts
@@ -1,3 +1,0 @@
-export { default as GithubIcon } from "./GithubIcon";
-export { default as LinkedinIcon } from "./LinkedinIcon";
-export { default as TwitterIcon } from "./TwitterIcon";
