+++ src/components/index.ts
@@ -1,5 +1,10 @@
-export * from "./content";
-export * from "./icons";
-export * from "./seo";
-export * from "./typography";
-export { default as ViewCounter, View } from "./ViewCounter";
+export { default as Analytics } from "./Analytics";
+export { default as GithubIcon } from "./GithubIcon";
+export { default as Heading } from "./Heading";
+export { default as HeadTags } from "./HeadTags";
+export { default as LinkedinIcon } from "./LinkedinIcon";
+export { default as Mdx } from "./Mdx";
+export { default as OGTags } from "./OGTags";
+export { default as Text } from "./Text";
+export { default as TwitterIcon } from "./TwitterIcon";
+export { default as Views } from "./Views";
