--- src/components/seo/index.ts
@@ -1,3 +1,0 @@
-export { default as Analytics } from "./Analytics";
-export { default as HeadTags } from "./HeadTags";
-export { default as OGTags } from "./OGTags";
