--- src/components/typography/index.ts
@@ -1,2 +1,0 @@
-export { default as Heading } from "./Heading";
-export { default as Text } from "./Text";
